Moved static methods to a new class JSONCommonUtils
